MB-19605: Fix compile warning in checked_snprintf
    checked_snprintf.cc:47:20: warning: comparison between signed and
    unsigned integer expressions [-Wsign-compare]

    } else if (ret >= size) {
                   ^

Reviewed-on: http://review.couchbase.org/64334
Tested-by: buildbot <build@couchbase.com>

+++ src/checked_snprintf.cc
@@ -44,7 +44,7 @@ int checked_snprintf(char* str, const size_t size, const char* format, ...) {
         // overflow error...
         throw std::overflow_error(
             "checked_snprintf: Destination buffer too small.");
-    } else if (ret >= size) {
+    } else if (size_t(ret) >= size) {
         std::string msg("checked_snprintf: Destination buffer too small. (");
         msg += std::to_string(ret);
         msg += " >= ";
